Add feature to enable packed headers.
While this may eventually be the default, right now it is possible
that there is still a safety issue as the there is a lot of
unsafe code in these headers.

This reduces the space usage for headers down to 1 word, which can
be useful for certain algorithms, as GC headers can now fit in the
same space as an Ix.